+++ rustsat/src/encodings/card/totalizer.rs
@@ -40,7 +40,7 @@ pub struct Totalizer {
     /// The root of the tree, if constructed
     root: Option<Node>,
     /// The number of variables in the totalizer
-    n_vars: usize,
+    n_vars: u32,
     /// The number of clauses in the totalizer
     n_clauses: usize,
 }
@@ -245,7 +245,7 @@ impl EncodeStats for Totalizer {
         self.n_clauses
     }
 
-    fn n_vars(&self) -> usize {
+    fn n_vars(&self) -> u32 {
         self.n_vars
     }
 }
